requirements.txt
numpy==1.12.1
olefile==0.44
Pillow==4.1.1
protobuf==3.2.0
scipy==0.19.0
six==1.10.0
tensorflow==1.1.0
Werkzeug==0.12.1
